03_RAG/requirements.txt
langchain==0.3.23
langchain-openai==0.3.12
langchain_community==0.3.21
langchain-chroma==0.2.3
chromadb==1.0.5
fastapi==0.115.12
uvicorn==0.34.0
gradio==5.24.0
concurrent-log-handler==0.9.25
pdfminer.six
nltk==3.9.1
sentence-transformers==3.4.1